Made Grid Search as the n. of p.c. increase
I made a similar thing as giuliano did: as the number of principal
components increase, I try to find the best parameter k for kNN using a
Grid Search. Results were really similar to the ones from giuliano, and
my approach requires a ton of time more, so maybe it's not worth it.